# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AWS Image Service API
  description: |
    Serverless image upload and AI analysis service built on AWS.

    ## Authentication
    Most endpoints require JWT authentication via AWS Cognito. Include the access token
    in the Authorization header as a Bearer token.

    ## User Roles
    - **user**: Can only access their own images and analysis results
    - **admin**: Can access all users' data, with optional `?userId=` filter
  version: 1.0.0
  contact:
    name: API Support
    email: john.gambrell@gmail.com
servers:
  - url: https://{api-id}.execute-api.{region}.amazonaws.com/prod
    description: Production API Gateway
    variables:
      api-id:
        default: xxxxxxxxxx
        description: API Gateway ID
      region:
        default: us-east-1
        description: AWS Region
...
components:
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT access token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
paths:
...
  /api/images/{imageId}:
    get:
      tags:
        - Images
      summary: Get image file
      description: |
        Download an image file. Returns a redirect to a presigned S3 URL.
        Users can only access their own images unless they are admin.
      operationId: getImage
      security:
        - CognitoAuth: []
      parameters:
        - name: imageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "302":
          description: Redirect to presigned S3 URL
          headers:
            Location:
              description: Presigned S3 URL for image download
              schema:
                type: string
                format: uri
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied - not owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...